Using IniConfigManager instead of duplicate
