+++ osu.Framework/Configuration/LocalSettingsManager.cs
@@ -1,13 +1,21 @@
 ﻿// Copyright (c) 2007-2018 ppy Pty Ltd <contact@ppy.sh>.
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu-framework/master/LICENCE
 
+using osu.Framework.Platform;
 using System;
 using System.IO;
 
 namespace osu.Framework.Configuration
 {
-    public class LocalSettingsManager : LocalIniConfigManager<LocalSetting>
+    public class LocalSettingsManager : IniConfigManager<LocalSetting>
     {
+        protected override string Filename => @"settings.ini";
+
+        public LocalSettingsManager(Storage storage)
+            : base(storage)
+        {
+        }
+
         protected override void InitialiseDefaults()
         {
             switch (RuntimeInfo.OS)
